feat: add timestamp

+++ philo/includes/philo.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   philo.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tkafanov <tkafanov@student.42vienna.com    +#+  +:+       +#+        */
+/*   By: tkafanov <tkafanov@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/16 14:16:46 by tkafanov          #+#    #+#             */
-/*   Updated: 2024/07/25 17:09:45 by tkafanov         ###   ########.fr       */
+/*   Updated: 2024/10/03 09:33:59 by tkafanov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -69,7 +69,7 @@ typedef struct s_data
 	int 			time_to_eat;
 	int 			time_to_sleep;
 	int				number_of_meals;
-	int				timestamp;
+	long			timestamp;
 	pthread_t		*threads;
 	pthread_mutex_t *fork;
 }	t_data;
